from instead of use

+++ Database/Database.hs
@@ -21,28 +21,28 @@ workWithDatabase :: String -> IO ()
 workWithDatabase name = do
     putStr $ name ++ " => "
     args <- split ' ' <$> getLine
-    case (head args) of
-        "create" -> do
-            fileExist <- doesFileExist $ toPath name $ head $ tail args
-            if fileExist then putStrLn "This table already exists!"
-                else if areTypes $ map (toTuple . split ':') $ split ',' $ last args
-                   then createTable name (head $ tail args) $ join $ tail $ tail args
-                   else putStrLn "Invalid declaration!"
-            workWithDatabase name
-        "drop" -> do
-            fileExist <- doesFileExist $ toPath name $ head $ tail args
-            if fileExist
-                then removeFile (toPath name $ head $ tail args)
-                  >> putStrLn ("Deleted table \"" ++ (head (tail args)) ++ "\"!")
-                else putStrLn "Selected table doesn't exist!"
-            (workWithDatabase name)
-        "use" -> do
-            fileExist <- doesFileExist $ toPath name $ head $ tail args
-            if fileExist then workWithTable name $ head $ tail args
-              else putStrLn "Selected table doesn't exist!"
-            (workWithDatabase name)
-        "ls" -> listDirectory ("./.databases/" ++ name) 
-            >>= print . map (join . init . split '.') >> (workWithDatabase name)
-        "exit" -> putStrLn "Exiting..."
-        "" -> workWithDatabase name
-        _      -> putStrLn "Invalid command!" >> (workWithDatabase name)
+    if (head $ split '#' $ last args) == "from" 
+       then do
+           fileExist <- doesFileExist $ toPath name $ last $ split '#' $ last args
+           if fileExist then workWithTable name (last $ split '#' $ last args) $ init args
+                        else putStrLn "This table doesn't exist!"
+       else case (head args) of
+            "create" -> do
+                fileExist <- doesFileExist $ toPath name $ head $ tail args
+                if fileExist then putStrLn "This table already exists!"
+                    else if areTypes $ map (toTuple . split ':') $ split ',' $ last args
+                       then createTable name (head $ tail args) $ join $ tail $ tail args
+                       else putStrLn "Invalid declaration!"
+                workWithDatabase name
+            "drop" -> do
+                fileExist <- doesFileExist $ toPath name $ head $ tail args
+                if fileExist
+                    then removeFile (toPath name $ head $ tail args)
+                      >> putStrLn ("Deleted table \"" ++ (head (tail args)) ++ "\"!")
+                    else putStrLn "Selected table doesn't exist!"
+                (workWithDatabase name)
+            "ls" -> listDirectory ("./.databases/" ++ name) 
+                >>= print . map (join . init . split '.') >> (workWithDatabase name)
+            "exit" -> putStrLn "Exiting..."
+            "" -> workWithDatabase name
+            _      -> putStrLn "Invalid command!" >> (workWithDatabase name)
